fixes for kernel 5

+++ scripts/copy_load.sh
@@ -15,18 +15,18 @@ then
     # sed -i "i/pass=$pass/" $0 # TODO: save to file
 fi
 
-echo "-------------------- router module files:"
-sshpass -p $pass ssh $host 'ls /lib/modules/ -tphl | grep -v /'
-echo "-------------------- router firmware files:"
-sshpass -p $pass ssh $host 'ls /lib/firmware/ -tphl | grep -v /'
+echo -e "\n-------------------- router firmware files:"
+sshpass -p $pass ssh $host 'ls -tphl /lib/firmware/ProgModel_* /lib/firmware/contr_lm.bin /lib/firmware/ap_upper_wave300.bin 2> /dev/null'
+echo -e "\n-------------------- router module files: (auto rsync)"
+sshpass -p $pass ssh $host 'ls -tphl /lib/modules/ | grep -v /'
 
-echo '--------------------  What do you want to do?';
+echo -e '\n-------------------- What do you want to do?';
 select option in 'Reboot the router' 'Copy and test modules and firmwares' 'Clear previus firmwares attempts' 'Start the driver stored on the router' 'Config it to run on boot'
 do
     if [ "$option" = "Reboot the router" ]
     then
         sshpass -p $pass ssh $host "reboot"
-        echo "wait lan reconnect before procede ..."
+        echo "... Wait for lan reconnection before procede ..."
     fi
 
     if [ "$option" = "Copy and test modules and firmwares" ]
@@ -39,9 +39,9 @@ do
     then
         for firmware in cal_wlan0 ap_upper_wave300 sta_upper_wave300 contr_lm ProgModel_BG_CB_wave300 ProgModel_BG_CB_3 ProgModel_BG_nCB_wave300 ProgModel_BG_nCB_3
         do
-            echo "Removing $firmware local link and router file"
-            rm $firmware
-            sshpass -p $pass ssh $host "rm /lib/firmware/$firmware"
+            echo "Removing local link and router file: $firmware*.bin"
+            [ -L $firmware*.bin ] && rm $firmware*.bin
+            sshpass -p $pass ssh $host "[ -f /lib/firmware/$firmware*.bin ] && rm /lib/firmware/$firmware*.bin"
         done    
         break;
     fi  
@@ -102,11 +102,11 @@ then
     #sshpass -p $pass ssh $host 'ln -s /lib/modules/key-build.pub /etc/opkg/keys/$(usign -F -p /lib/modules/key-build.pub)'
     #sshpass -p $pass ssh $host "sed -i 's/downloads.openwrt.org\/releases\/19.07.7/$( ifconfig | sed -En 's/127.0.0.1//;s/.*inet (addr:)?(([0-9]*\.){3}[0-9]*).*/\2/p' )/' /etc/opkg/distfeeds.conf"
     #sshpass -p $pass ssh $host "sed -i 's/$( ifconfig | sed -En 's/127.0.0.1//;s/.*inet (addr:)?(([0-9]*\.){3}[0-9]*).*/\2/p' )/downloads.openwrt.org\/releases\/19.07.7/' /etc/opkg/distfeeds.conf"
-    
-    #echo "Updating and installing some tools ..."
-    sshpass -p $pass ssh $host 'opkg update; opkg install rsync'
+
+    echo "Updating and installing some tools ..."
+    sshpass -p $pass ssh $host 'opkg update; opkg install rsync libopenssl'
     #sshpass -p $pass ssh $host 'opkg install kmod-usb-storage block-mount kmod-fs-vfat kmod-nls-cp437 kmod-nls-iso8859-1 '
-    #sshpass -p $pass ssh $host 'opkg install pciutils wpad wireless-tools iw iwinfo libopenssl' # not installing from local binaries
+    #sshpass -p $pass ssh $host 'opkg install pciutils wpad wireless-tools iw iwinfo ' # not installing from local binaries
     #sshpass -p $pass ssh $host 'opkg install libnl libnl-tiny' #    kmod-lib80211 kmod-mac80211 kmod-cfg80211'
 fi
 
@@ -114,11 +114,11 @@ fi
 echo "Coping/Updating modules to the router ..."
 sshpass -p $pass rsync -caL config.conf mtlkroot.ko mtlk.ko mtlk-ap $host:/lib/modules/
 
-echo "-------------------- Starting mtlkroot.ko ..."
+echo "-------------------- starting mtlkroot.ko ..."
 sshpass -p $pass ssh $host 'insmod /lib/modules/mtlkroot.ko cdebug=3'
-echo "-------------------- Starting mtlk.ko ..."
-sshpass -p $pass ssh $host 'insmod /lib/modules/mtlk.ko ap=1; dmesg | tail -n 69'
-
+# here should check if mtlkroot.ko loaded correctly
+echo "-------------------- insmod mtlk.ko ..."
+sshpass -p $pass ssh $host 'insmod /lib/modules/mtlk.ko ap=1'
 
 # - cal_wlan0 calibration data -> probably on EEPROM
 # ap_upper_wave300 -> access point operation
@@ -126,49 +126,47 @@ sshpass -p $pass ssh $host 'insmod /lib/modules/mtlk.ko ap=1; dmesg | tail -n 69
 # contr_lm.bin: Lower interface (PCI/PCIe communication with the driver)
 # ProgModel_BG_nCB_wave300.bin - PHY firmware (?)
 # ProgModel_BG_nCB_3D_RevB_wave300.bin - HW firmware (?)
-text="--------------------\nSelect the new firmware cadidate, or \n '1' to procede to the next firmware, or \n any other value, removes from router and scp the candidate: "
+IFS=$'\n'
 cd ../firmware
 for firmware in cal_wlan0 ap_upper_wave300 sta_upper_wave300 contr_lm ProgModel_BG_CB_wave300 ProgModel_BG_CB_3 ProgModel_BG_nCB_wave300 ProgModel_BG_nCB_3
 do
-    files="Next $(find -name "$firmware*.bin")"
-    if [[ "$files" != "Next " ]]
+    files=$(find -name "$firmware*.bin")
+    if [[ "$files" != "" ]] && ! [ -L ../scp/$firmware*.bin ]
     then
-        last=$(ls ../scp/$firmware*.bin -t1 2> /dev/null | head -n 1 ) #; echo $last
-        if [ "$firmware" = 'ProgModel_BG_CB_wave300' ]
-        then
-            sshpass -p $pass ssh $host 'brctl addif "br-lan" "wlan0" ; iwlist wlan0 f ; iwinfo ; ifconfig wlan0 ; brctl show' # bridge and show info
-            sshpass -p $pass ssh $host '/lib/modules/mtlk-ap -d /lib/modules/config.conf ; dmesg | tail -n 69' # load hostapd
-        fi
-        echo -e "$text$last"
-        select file in $files
+        [ "$firmware" = 'ProgModel_BG_CB_wave300' ] && echo "-------------------- Configuring bridge" && sshpass -p $pass ssh $host 'brctl addif "br-lan" "wlan0" ; iwlist wlan0 f ; iwinfo ; ifconfig wlan0 ; brctl show'
+
+        while true 
         do
-            if [ "$file" = "Next" ]
-            then
-                break
-            fi
-            
-            if [[ "$last" != "" ]] 
-            then
-                rm ../scp/$(basename $last)
-                sshpass -p $pass ssh $host "rm /lib/firmware/$(basename $last)" # remove previus firmware
-            fi
-            last=$file
-            
-            if [[ "$file" != "" ]] 
-            then
-                ln -s $file ../scp/$(basename $file) # ?
-                sshpass -p $pass scp $file $host:/lib/firmware/ # copy new candidate
+            sshpass -p $pass ssh $host 'dmesg | tail -n 69' | grep --color=auto -nw $(basename $firmware)
+            echo -e "\n-------------------- What do you want to try?"
+            last="Last file worked, so lets procede to the next"
+            select file in $files
+            do
+                [ "$file" = $last ] && break;
+
+                [ -L ../scp/$firmware*.bin ] && rm ../scp/$firmware*.bin # remove previus firmware
+                sshpass -p $pass ssh $host "[ -f /lib/firmware/$firmware*.bin ] && rm /lib/firmware/$firmware*.bin"
+
+                [ "$file" = "" ] && exit;
+
+                echo -e "Creating link and coping file to the router ..."
+                ln -s $file ../scp/$(basename $file)
+                sshpass -p $pass scp $file $host:/lib/firmware/
                 if [[ "${firmware:0:9}" != "ProgModel" ]]
                 then
-                    echo "-------------------- loading mtlk.ko ..."
-                    sshpass -p $pass ssh $host 'rmmod /lib/modules/mtlk.ko' # ?
-                    sshpass -p $pass ssh $host 'insmod /lib/modules/mtlk.ko ap=1; dmesg | tail -n 69' # try load wave300
+                    echo "-------------------- insmod mtlk.ko ..."
+#                    sshpass -p $pass ssh $host 'rmmod /lib/modules/mtlk.ko' # crashes the router
+                    sshpass -p $pass ssh $host 'insmod /lib/modules/mtlk.ko ap=1' # try insmod wave300
+#                   [ $? = 0 ] && file=$last # if no errors, then go to the next
                 else
                     echo "-------------------- starting hostapd ..."
-                    sshpass -p $pass ssh $host '/lib/modules/mtlk-ap -d /lib/modules/config.conf ; dmesg | tail -n 69' # try start hostapd
+                    sshpass -p $pass ssh $host '/lib/modules/mtlk-ap -d /lib/modules/config.conf' # try start hostapd
+#                   [ $? = 0 ] && exit;
                 fi
-            fi
-            echo -e "$text$last"
+                break;
+            done
+        [ "${files:0:9}" = "${last:0:9}" ] || files=$last$'\n'$files
+        [ "$file" = "$last" ] && break;
         done
     fi
 done
